Multiply width by height using the int64_t type
Also perform some other multiplications using the int64_t type. Avoid
integer overflows when width and height are large (e.g., 1 << 15 or
1 << 16).

Bug: aomedia:3508

+++ av1/encoder/ratectrl.c
@@ -334,7 +334,7 @@ int av1_rc_get_default_min_gf_interval(int width, int height,
                                        double framerate) {
   // Assume we do not need any constraint lower than 4K 20 fps
   static const double factor_safe = 3840 * 2160 * 20.0;
-  const double factor = width * height * framerate;
+  const double factor = (double)width * height * framerate;
   const int default_interval =
       clamp((int)(framerate * 0.125), MIN_GF_INTERVAL, MAX_GF_INTERVAL);
